requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0
langgraph>=0.1.0

duckduckgo-search>=6.3.0
ddgs>=1.1.6
wikipedia>=1.4.0
requests>=2.31.0

chromadb>=0.5.5        
python-dotenv>=1.0.1
pydantic>=2.8.2
tqdm>=4.66.1
asknews>=0.12.5
arxiv>=2.2.0
sympy>=1.12
rich>=13.7.1